Set `engine.vscode` in `package.json` Accidentally left out of https://github.com/microsoft/vscode-python/commit/cb124e7b408af2afa3e074e541ddf8c09b95e2ff .
As it pertains to a change made in VS Code insiders today, we are setting the engine to a specific date. This requires updating the `getChannel()` code, so took the opportunity to factor it out.
